Fixed typing of Spell/aura effects. Fixed typing of damage/HP Damage effects are floats in the DBC and they were getting rounded to ints.
The caused a mis-match when calculating damge, shields, HP and healing.

All types made floats/long for maximum compatability. More work is required on types.

+++ Source/Game/AI/CoreAI/PetAI.cs
@@ -641,7 +641,7 @@ namespace Game.AI
             }
         }
 
-        public override void DamageTaken(Unit attacker, ref uint damage, DamageEffectType damageType, SpellInfo spellInfo = null)
+        public override void DamageTaken(Unit attacker, ref float damage, DamageEffectType damageType, SpellInfo spellInfo = null)
         {
             AttackStart(attacker);
         }
