+++ Source/Game/AI/CoreAI/UnitAI.cs
@@ -514,10 +514,10 @@ namespace Game.AI
         public virtual void OnDespawn() { }
 
         // Called at any Damage to any victim (before damage apply)
-        public virtual void DamageDealt(Unit victim, ref uint damage, DamageEffectType damageType) { }
-        public virtual void DamageTaken(Unit attacker, ref uint damage, DamageEffectType damageType, SpellInfo spellInfo = null) { }
-        public virtual void HealReceived(Unit by, uint addhealth) { }
-        public virtual void HealDone(Unit to, uint addhealth) { }
+        public virtual void DamageDealt(Unit victim, ref float damage, DamageEffectType damageType) { }
+        public virtual void DamageTaken(Unit attacker, ref float damage, DamageEffectType damageType, SpellInfo spellInfo = null) { }
+        public virtual void HealReceived(Unit by, float addhealth) { }
+        public virtual void HealDone(Unit to, float addhealth) { }
         public virtual void SpellInterrupted(uint spellId, uint unTimeMs) { }
 
         /// <summary>
