+++ Source/Scripts/EasternKingdoms/Karazhan/Opera.cs
@@ -1095,7 +1095,7 @@ namespace Scripts.EasternKingdoms.Karazhan.EsOpera
             }
         }
 
-        public override void DamageTaken(Unit done_by, ref uint damage, DamageEffectType damageType, SpellInfo spellInfo = null)
+        public override void DamageTaken(Unit done_by, ref float damage, DamageEffectType damageType, SpellInfo spellInfo = null)
         {
             if (damage < me.GetHealth())
                 return;
@@ -1399,7 +1399,7 @@ namespace Scripts.EasternKingdoms.Karazhan.EsOpera
             me.DespawnOrUnsummon();
         }
 
-        public override void DamageTaken(Unit done_by, ref uint damage, DamageEffectType damageType, SpellInfo spellInfo = null)
+        public override void DamageTaken(Unit done_by, ref float damage, DamageEffectType damageType, SpellInfo spellInfo = null)
         {
             if (damage < me.GetHealth())
                 return;
